+++ db/migration00005.sql
@@ -1,1 +1,4 @@
+ALTER TABLE masterversions ADD devlastcommit timestamp;
+ALTER TABLE masterversions ADD releasedate timestamp;
+ALTER TABLE pods ADD daysmonitored int, ADD countryname text;
 ALTER TABLE pods ADD detectedlanguage text;