+++ db/tables.sql
@@ -14,6 +14,7 @@ CREATE TABLE pods (
  ip text,
  detectedlanguage text,
  country text,
+ countryname text,
  city text,
  state text, 
  lat text,
@@ -23,6 +24,7 @@ CREATE TABLE pods (
  secure boolean,
  sslvalid text,
  monthsmonitored int,
+ daysmonitored int,
  signup boolean,
  total_users int, 
  active_users_halfyear int,
@@ -95,6 +97,8 @@ CREATE TABLE masterversions (
  id serial8 UNIQUE PRIMARY KEY,
  software text,
  version text,
+ devlastcommit timestamp,
+ releasedate timestamp,
  date_checked timestamp DEFAULT current_timestamp
 );